File I/O included
Previously, the monitored processes were stored in memory, such that when a process that has started being monitored falls below the memory threshold, the Process Monitor will still continue monitoring this process. The problem was that a restart of the machine agent caused a loss of the set of processes that are monitored. In this version of the Process Monitor, the set of monitored processes is stored in a file. Upon restart of the machine-agent, there will no longer be a loss of required information